requirements.txt
urllib3==1.26.14
Pillow==9.4.0
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Flask==2.2.2
Flask-Cors==3.0.10
tensorflow==2.10.0
tensorboard==2.10.1
uuid==0.2.7
plot-keras-history==1.1.38
seaborn==0.12.2
matplotlib==3.5.3
opencv-python==4.7.0.68
pandas==1.1.5
jupyter_client==7.4.9
jupyter_core==4.12.0
numpy==1.21.6